+++ src/components/Requirement/RequirementControlBtn.tsx
@@ -10,7 +10,7 @@ import { theme, ThemeType } from '@/styles/theme';
 import { RequirementStateType } from '@/types/requirement';
 import { copyText, generateSearchPrompt } from '@/utils/gpt';
 
-import DeleteHoverBtn from '../Shared/DeleteHoverBtn';
+import DeleteHoverBtn from '../Shared/Button/DeleteHoverBtn';
 import HoverIcon from '../Shared/HoverIcon';
 import GPTGenerateCodeBtn from '../gpt/GPTGenerateCodeBtn';
 import SearchIcon from '../icons/SearchIcon';
